Start work on regression in TPOT
