+++ tpot/operators/regressors/__init__.py
@@ -12,15 +12,14 @@ any later version.
 
 The TPOT library is distributed in the hope that it will be useful, but
 WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
-You should have received a copy of the GNU General Public License along with
-the TPOT library. If not, see http://www.gnu.org/licenses/.
+FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+details. You should have received a copy of the GNU General Public License along
+with the TPOT library. If not, see http://www.gnu.org/licenses/.
 
 """
 
 from .base import *
-from .classifiers import *
-from .preprocessors import *
-from .selectors import *
-from .regressors import *
-from .combine_dfs import CombineDFs
+from .extra_trees import *
+from .random_forest import *
+from .knnr import *
+from .gradient_boosting import *
