+++ tpot/operators/regressors/extra_trees.py
@@ -18,43 +18,30 @@ with the TPOT library. If not, see http://www.gnu.org/licenses/.
 
 """
 
-from .base import Selector
-from sklearn.feature_selection import SelectFromModel
-from sklearn.ensemble import ExtraTreesClassifier
+from .base import Regressor
+from sklearn.ensemble import ExtraTreesRegressor
 
 
-class TPOTSelectFromModel(Selector):
-    """Uses scikit-learn's ExtraTreesClassifier combined with SelectFromModel
-    to transform the feature set.
+class TPOTExtraTreesRegressor(Regressor):
+    """Fits an Extra Trees Regressor
 
     Parameters
     ----------
-    threshold: float
-        Features whose importance is greater or equal are kept while the others
-        are discarded.
     criterion: int
-        For the ExtraTreesClassifier:
         Integer that is used to select from the list of valid criteria,
         either 'gini', or 'entropy'
     max_features: float
-        For the ExtraTreesClassifier:
         The number of features to consider when looking for the best split
 
     """
-    import_hash = {
-        'sklearn.feature_selection': ['SelectFromModel'],
-        'sklearn.ensemble':          ['ExtraTreesClassifier']
-    }
-    sklearn_class = SelectFromModel
-    arg_types = (float, int, float)
-    regression = False  # Can not be used in regression due to ExtraTreesClassifier
+    import_hash = {'sklearn.ensemble': ['ExtraTreesRegressor']}
+    sklearn_class = ExtraTreesRegressor
+    arg_types = (int, float)
 
     def __init__(self):
         pass
 
-    def preprocess_args(self, threshold, criterion, max_features):
-        threshold = min(1., max(0., threshold))
-
+    def preprocess_args(self, criterion, max_features):
         # Select criterion string from list of valid parameters
         criterion_values = ['gini', 'entropy']
         criterion_selection = criterion_values[criterion % len(criterion_values)]
@@ -62,6 +49,7 @@ class TPOTSelectFromModel(Selector):
         max_features = min(1., max(0., max_features))
 
         return {
-            'estimator': ExtraTreesClassifier(criterion=criterion_selection, max_features=max_features),
-            'threshold': threshold
+            'criterion': criterion_selection,
+            'max_features': max_features,
+            'n_estimators': 500
         }
